chore: move interactive stuff to its own package
