+++ cmd/api.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/speakeasy-api/speakeasy/internal/api"
-	"github.com/speakeasy-api/speakeasy/internal/utils"
+	"github.com/speakeasy-api/speakeasy/internal/interactivity"
 	"github.com/spf13/cobra"
 )
 
@@ -12,7 +12,7 @@ var apiCmd = &cobra.Command{
 	Long: `Provides access to the Speakeasy API via the CLI.
 To authenticate with the Speakeasy API, you must first create an API key via https://app.speakeasyapi.dev
 and then set the SPEAKEASY_API_KEY environment variable to the value of the API key.`,
-	RunE: utils.InteractiveRunFn("Choose an API endpoint:"),
+	RunE: interactivity.InteractiveRunFn("Choose an API endpoint:"),
 }
 
 func apiInit() {
