+++ cmd/auth.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/speakeasy-api/speakeasy/internal/auth"
-	"github.com/speakeasy-api/speakeasy/internal/utils"
+	"github.com/speakeasy-api/speakeasy/internal/interactivity"
 	"github.com/spf13/cobra"
 )
 
@@ -10,7 +10,7 @@ var authCmd = &cobra.Command{
 	Use:   "auth",
 	Short: "Authenticate the CLI",
 	Long:  `The "authenticate" command allows control over the authentication of the CLI.`,
-	RunE:  utils.InteractiveRunFn("Choose an option:"),
+	RunE:  interactivity.InteractiveRunFn("Choose an option:"),
 }
 
 var loginCmd = &cobra.Command{
