+++ cmd/overlay.go
@@ -1,13 +1,13 @@
 package cmd
 
 import (
+	"github.com/speakeasy-api/speakeasy/internal/interactivity"
 	"github.com/speakeasy-api/speakeasy/internal/log"
 	"os"
 
 	"github.com/spf13/cobra"
 
 	"github.com/speakeasy-api/speakeasy/internal/overlay"
-	"github.com/speakeasy-api/speakeasy/internal/utils"
 )
 
 var overlayCmd = &cobra.Command{
@@ -19,7 +19,7 @@ var overlayValidateCmd = &cobra.Command{
 	Use:     "validate",
 	Short:   "Given an overlay, it will state whether it appears to be valid according to the OpenAPI Overlay specification",
 	Args:    cobra.NoArgs,
-	PreRunE: utils.GetMissingFlagsPreRun,
+	PreRunE: interactivity.GetMissingFlagsPreRun,
 	RunE:    runValidateOverlay,
 }
 
@@ -27,7 +27,7 @@ var overlayCompareCmd = &cobra.Command{
 	Use:     "compare",
 	Short:   "Given two specs, it will output an overlay that describes the differences between them",
 	Args:    cobra.NoArgs,
-	PreRunE: utils.GetMissingFlagsPreRun,
+	PreRunE: interactivity.GetMissingFlagsPreRun,
 	RunE:    runCompare,
 }
 
@@ -35,7 +35,7 @@ var overlayApplyCmd = &cobra.Command{
 	Use:     "apply",
 	Short:   "Given an overlay, it will construct a new specification by extending a specification and applying the overlay, and output it to stdout.",
 	Args:    cobra.NoArgs,
-	PreRunE: utils.GetMissingFlagsPreRun,
+	PreRunE: interactivity.GetMissingFlagsPreRun,
 	RunE:    runApply,
 }
