+++ internal/interactivity/button.go
@@ -1,4 +1,4 @@
-package utils
+package interactivity
 
 import (
 	"github.com/charmbracelet/lipgloss"
