+++ internal/interactivity/multiInput.go
@@ -1,4 +1,4 @@
-package utils
+package interactivity
 
 import (
 	"fmt"
